Modify file selection algorithm
This commit changes the file selection algorithm so that duplicate files
no longer count toward the maximum size specified by the user. The size
argument should reflect the amount of data that the user is willing to
not longer have access to. When duplicates of files are removed, the
originals are still accessible. That is the rationale for the change.
There is now an option to disable the automatic inclusion of duplicate
files.