Change handling of duplicate files
Previously, the program would suggest the most recently-modified files
from each group of duplicate files. Now, it suggests all but the most
recently modified file from each group of duplicate files. The rationale
for this change is that the user should only need one copy of each of
their files.